+++ service/controller/resource/prometheus/resource_test.go
@@ -23,6 +23,7 @@ func TestPrometheus(t *testing.T) {
 		Installation:      "test-installation",
 		Pipeline:          "testing",
 		Provider:          "provider",
+		PrometheusVersion: "v2.24.0",
 		Region:            "onprem",
 		Registry:          "quay.io",
 		StorageSize:       "50Gi",