+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix recording rules creation
+- Fix prometheus container image tag to not use latest
 
 ## [1.14.0] - 2021-01-13